Add usage collection for queries
This changelist adds usage collection for queries. The collection is done
at the Cursor level. The hope is that collecting at this level, there will
be less code changes, and more uniformed code changes.

A UsageCollector is added to each Query type to store necessary information
to collect usage. This UsageCollector is passed to Cursor, so that the
collection can be done here.